修改:         software1/md5/1234.txt 	修改:         software1/md5/main.c 	修改:         software1/md5/md5.c

+++ software1/md5/md5.c
@@ -1,6 +1,6 @@
 #include <string.h>  
 
-typedef unsigned long UINT4;  
+typedef unsigned int UINT4;  
 typedef unsigned char *POINTER;  
  
 #define S11 7
